% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angenieux.R
\name{Angenieux}
\alias{Angenieux}
\title{R6 Class for visualising floundeR based datasets}
\description{
This class aims to provide aplanat-like visualisation abstraction for the
floundeR framework
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colourMap}}{\itemize{
\item default colourMap for plots requiring discrete colours
}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{A method to dump out the stored data from an \code{Angenieux} object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Angenieux-new}{\code{Angenieux$new()}}
\item \href{#method-Angenieux-add}{\code{Angenieux$add()}}
\item \href{#method-Angenieux-plot}{\code{Angenieux$plot()}}
\item \href{#method-Angenieux-to_file}{\code{Angenieux$to_file()}}
\item \href{#method-Angenieux-set_title}{\code{Angenieux$set_title()}}
\item \href{#method-Angenieux-clone}{\code{Angenieux$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Angenieux-new"></a>}}
\if{latex}{\out{\hypertarget{method-Angenieux-new}{}}}
\subsection{Method \code{new()}}{
Creates a new Angenieux object. This
initialisation method performs other sanity checking of the defined
file(s) and creates the required data structures.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$new(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{the datatype that is being passed to the object}

\item{\code{value}}{the data that is being passed to the object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Angenieux} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Angenieux-add"></a>}}
\if{latex}{\out{\hypertarget{method-Angenieux-add}{}}}
\subsection{Method \code{add()}}{
add an \code{AngenieuxDecoration} to the plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$add(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{an \code{AngenieuxDecoration}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Angenieux-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Angenieux-plot}{}}}
\subsection{Method \code{plot()}}{
Prepare and present an Angenieux plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters passed on to downstream methods - please see
examples for further examples as to how Angenieux plots can be customised
using this approach.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Angenieux-to_file"></a>}}
\if{latex}{\out{\hypertarget{method-Angenieux-to_file}{}}}
\subsection{Method \code{to_file()}}{
Specify that Angenieux plot should be saved to file

When working at the console an Angenieux plot may be plotted directly
to the console. When preparing reports through Rmarkdown or Pkgdown a
more logical saving of plots to a discrete file location may make more
sense. The method is used to instruct Angenieux that the plot should be
saved to a given location and with a given file format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$to_file(
  target_file,
  width = 18,
  height = 12,
  units = "cm",
  dpi = "print"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{target_file}}{the file with extension e.g. \code{figure1.png}}

\item{\code{width}}{the width of figure to save (12 by default)}

\item{\code{height}}{the height of figure to save (7.5 by default)}

\item{\code{units}}{the unit to use for height and width (cm by default)}

\item{\code{dpi}}{the plot resolution (print/300 by default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the original Angenieux object (self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Angenieux-set_title"></a>}}
\if{latex}{\out{\hypertarget{method-Angenieux-set_title}{}}}
\subsection{Method \code{set_title()}}{
Set the title used in the given Angenieux plot
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$set_title(title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{\itemize{
\item the title to use on the plot
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
the original Angenieux object (self)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Angenieux-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Angenieux-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Angenieux$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
