% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundled_data.R
\name{get_bundled_path}
\alias{get_bundled_path}
\title{extract the path from a bundled file}
\usage{
get_bundled_path(file)
}
\arguments{
\item{file}{a known filename (or its prefix)}
}
\value{
a file
}
\description{
A collection of files are distributed with the rtqc package; this
accessory method is aimed to abbreviate the technical code for the
identification of these files and to clean the vignettes and code examples.
}
