% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_indexed_tuples}
\alias{get_indexed_tuples}
\title{get the information on what has already been parsed within the working folder}
\usage{
get_indexed_tuples(dir)
}
\description{
get the information on what has already been parsed within the working folder
}
