% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_mean_qscore}
\alias{get_mean_qscore}
\title{calculate mean quality score from a vector of phred scores}
\usage{
get_mean_qscore(phred_scores)
}
\description{
calculate mean quality score from a vector of phred scores
}
