% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_qscore}
\alias{get_qscore}
\title{calculate mean quality score from an ASCII quality string}
\usage{
get_qscore(qualstr)
}
\description{
calculate mean quality score from an ASCII quality string
}
