% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{index_fastq_list}
\alias{index_fastq_list}
\title{perform an index of multiple fastq entry metadata}
\usage{
index_fastq_list(file_list, dir, threads)
}
\description{
perform an index of multiple fastq entry metadata
}
