% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infographic.R
\name{Infographic}
\alias{Infographic}
\title{R6 Class for loading and analysing sequence sets}
\description{
R6 Class for loading and analysing sequence sets

R6 Class for loading and analysing sequence sets
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{panel.width}}{defines width of an infographic block}

\item{\code{panel.height}}{defines height of an infographic block}

\item{\code{panel.spacer}}{defines spacing around an infographic block}

\item{\code{panel.x.offset}}{defines x offset spacing for whole graphic}

\item{\code{panel.y.offset}}{defines y offset spacing for whole graphic}

\item{\code{columns}}{defines the number of columns to use in infographic
plot}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{items}}{return an integer describing the number of items that is contained
within the \code{Infographic}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Infographic-new}{\code{Infographic$new()}}
\item \href{#method-Infographic-add}{\code{Infographic$add()}}
\item \href{#method-Infographic-as_tibble}{\code{Infographic$as_tibble()}}
\item \href{#method-Infographic-plot}{\code{Infographic$plot()}}
\item \href{#method-Infographic-shinyplot}{\code{Infographic$shinyplot()}}
\item \href{#method-Infographic-display_fa}{\code{Infographic$display_fa()}}
\item \href{#method-Infographic-clone}{\code{Infographic$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-new"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance of the R6 Class \code{Infographic}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-add"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-add}{}}}
\subsection{Method \code{add()}}{
add an \code{InfographicItem} to the Infographic plot.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$add(item)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{item}}{an \code{InfographicItem}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-as_tibble"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-as_tibble}{}}}
\subsection{Method \code{as_tibble()}}{
Export the contained \code{Infographic} dataset(s) as a tibble
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$as_tibble()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A tibble representation for all the data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-plot"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-plot}{}}}
\subsection{Method \code{plot()}}{
Plot the infographic to file (and display it immediately)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$plot(
  display_file = tempfile(pattern = "file", tmpdir = tempdir(), fileext = ".png")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{display_file}}{the file to write to the infographic to (a temp
file will be created and used by default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-shinyplot"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-shinyplot}{}}}
\subsection{Method \code{shinyplot()}}{
Emit the infographic as a plot for rendering in e.g. shiny
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$shinyplot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-display_fa"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-display_fa}{}}}
\subsection{Method \code{display_fa()}}{
Display a collection of fontawesome based infographics for picking.

The \code{fontawesome} collection of icons contains over 700 icons of
which some are more useful / desirable than others. This accessory
method is used to render an Infographic report that summarised the
available icons within a predefined range - the intention here is to
make the selection of fonts to use in infographics a little simpler
and easier. This replaces a dodgy notebook approach that was used
previously.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$display_fa(file, offset = 0, rows = 10, columns = 6)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{\itemize{
\item a file.path to use to write the infographic to
}}

\item{\code{offset}}{\itemize{
\item an integer offset defining where we should start
rendering from in a broad sequence.
}}

\item{\code{rows}}{\itemize{
\item the number of rows to fill with sequential data.
}}

\item{\code{columns}}{\itemize{
\item the corresponding number of columns.
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Infographic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Infographic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Infographic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
