% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_methods.R
\name{n50calc}
\alias{n50calc}
\title{report the N50 (or other value) of sequence lengths from a length vector}
\usage{
n50calc(data, n = 0.5)
}
\arguments{
\item{data}{\itemize{
\item the vector of sequence lengths
}}

\item{n}{\itemize{
\item the N value to present (0.5 by default)
}}
}
\description{
Trivial method to report the N50 from a sequence collection
}
