% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_set_summary.R
\name{sequence_set_summary}
\alias{sequence_set_summary}
\title{R6 Class for reporting sequence data facets}
\description{
\code{sequence_set_summary} class provides tools and widgets for the presentation
of sequence_set information
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_set_summary-new}{\code{sequence_set_summary$new()}}
\item \href{#method-sequence_set_summary-shiny_touch}{\code{sequence_set_summary$shiny_touch()}}
\item \href{#method-sequence_set_summary-touch}{\code{sequence_set_summary$touch()}}
\item \href{#method-sequence_set_summary-get_data}{\code{sequence_set_summary$get_data()}}
\item \href{#method-sequence_set_summary-summary}{\code{sequence_set_summary$summary()}}
\item \href{#method-sequence_set_summary-read_count}{\code{sequence_set_summary$read_count()}}
\item \href{#method-sequence_set_summary-read_bases}{\code{sequence_set_summary$read_bases()}}
\item \href{#method-sequence_set_summary-quality_highlights}{\code{sequence_set_summary$quality_highlights()}}
\item \href{#method-sequence_set_summary-length_highlights}{\code{sequence_set_summary$length_highlights()}}
\item \href{#method-sequence_set_summary-temporal_highlights}{\code{sequence_set_summary$temporal_highlights()}}
\item \href{#method-sequence_set_summary-pore_highlights}{\code{sequence_set_summary$pore_highlights()}}
\item \href{#method-sequence_set_summary-clone}{\code{sequence_set_summary$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-new}{}}}
\subsection{Method \code{new()}}{
Constructor method for \code{sequence_set_summary} requires just a
sequence_set object that will provide the other file interaction
capabilities
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$new(path = NULL, sequence_set = NULL, threads = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sequence_set}}{an existing sequence_set object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-shiny_touch"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-shiny_touch}{}}}
\subsection{Method \code{shiny_touch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$shiny_touch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-touch"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-touch}{}}}
\subsection{Method \code{touch()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$touch()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-get_data}{}}}
\subsection{Method \code{get_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$get_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-summary"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-summary}{}}}
\subsection{Method \code{summary()}}{
summary aggregates several pieces of information into a single call
that are routinely shown in infographics and dashboards
\itemize{
\item number of sequence reads
\item number of bases read
\item number of barcodes
\item length min, max, mean, median
\item quality min, max, mean, median
\item flowcell id(s)
\item earliest read timestamp
\item latest read timestamp
\item runtime duration
\item active pores (as pores seen within e.g. 15 minutes of latest read)
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-read_count"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-read_count}{}}}
\subsection{Method \code{read_count()}}{
returns the number of sequence reads that are defined within the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$read_count()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-read_bases"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-read_bases}{}}}
\subsection{Method \code{read_bases()}}{
returns a numeric defining the number of bases observed
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$read_bases(scale = "Gb", dpoints = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scale}}{\itemize{
\item the scale to use when reporting the number of bases; this
can be Mb, Gb, Tb (at this moment) - or (auto) that will select an
appropriate scale
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-quality_highlights"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-quality_highlights}{}}}
\subsection{Method \code{quality_highlights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$quality_highlights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-length_highlights"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-length_highlights}{}}}
\subsection{Method \code{length_highlights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$length_highlights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-temporal_highlights"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-temporal_highlights}{}}}
\subsection{Method \code{temporal_highlights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$temporal_highlights(unit = "hours")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-pore_highlights"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-pore_highlights}{}}}
\subsection{Method \code{pore_highlights()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$pore_highlights()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set_summary-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set_summary-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set_summary$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
