% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basecalled_folder.R
\name{basecalled_folder}
\alias{basecalled_folder}
\title{R6 Class for handling folders of basecalled sequence from dorado or MinKNOW}
\description{
\code{basecalled_folder} class provides an infrastructure for working with
the output from sequence basecalling.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-basecalled_folder-new}{\code{basecalled_folder$new()}}
\item \href{#method-basecalled_folder-print}{\code{basecalled_folder$print()}}
\item \href{#method-basecalled_folder-index}{\code{basecalled_folder$index()}}
\item \href{#method-basecalled_folder-as_sequence_set}{\code{basecalled_folder$as_sequence_set()}}
\item \href{#method-basecalled_folder-status}{\code{basecalled_folder$status()}}
\item \href{#method-basecalled_folder-status_str}{\code{basecalled_folder$status_str()}}
\item \href{#method-basecalled_folder-get_cache_dir}{\code{basecalled_folder$get_cache_dir()}}
\item \href{#method-basecalled_folder-is_new_sequence_data}{\code{basecalled_folder$is_new_sequence_data()}}
\item \href{#method-basecalled_folder-get_sequence_file_info}{\code{basecalled_folder$get_sequence_file_info()}}
\item \href{#method-basecalled_folder-list_files}{\code{basecalled_folder$list_files()}}
\item \href{#method-basecalled_folder-clone}{\code{basecalled_folder$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-new"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-new}{}}}
\subsection{Method \code{new()}}{
constructor to initialise an instance of the basecalled_folder
class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$new(
  seq_path,
  use_unclassified = private$.use_unclassified,
  cache_dir = tempdir()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_path}}{\itemize{
\item the sequence path to investigate
}}

\item{\code{use_unclassified}}{\itemize{
\item when using demultiplexed datasets, include
the unclassified sequences as a valid "barcode"
}}

\item{\code{cache_dir}}{\itemize{
\item where should intermediate files be stored; this
will utilise a temporary directory by default
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-print"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-print}{}}}
\subsection{Method \code{print()}}{
this print method overrides the standard print function as included with
R6 objects - this is to better define what is contained within the object
and to provide better \code{rtqc} abstraction
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional stuff passed on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing (at present) - output to stdout
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-index"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-index}{}}}
\subsection{Method \code{index()}}{
the index method will index the available sequence files to recover
an object that is not entirely dissimilar to the original
sequencing_summary file that was produced by the guppy software
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$index(threads = 2)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threads}}{the number of threads to run in parallel}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-as_sequence_set"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-as_sequence_set}{}}}
\subsection{Method \code{as_sequence_set()}}{
prepares a \code{sequence_set} object from the current basecalled folder
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$as_sequence_set()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-status"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-status}{}}}
\subsection{Method \code{status()}}{
a public accessory method to check on what is happening with the
current \code{basecalled_folder}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$status(echo = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-status_str"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-status_str}{}}}
\subsection{Method \code{status_str()}}{
a public accessory method to check on what is happening with the
current \code{basecalled_folder}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$status_str()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-get_cache_dir"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-get_cache_dir}{}}}
\subsection{Method \code{get_cache_dir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$get_cache_dir()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-is_new_sequence_data"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-is_new_sequence_data}{}}}
\subsection{Method \code{is_new_sequence_data()}}{
check whether new sequences have become available within the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$is_new_sequence_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-get_sequence_file_info"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-get_sequence_file_info}{}}}
\subsection{Method \code{get_sequence_file_info()}}{
return a tibble containing information on the sequence files identified
within this \code{basecalled_folder} instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$get_sequence_file_info(qualified_path = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{qualified_path}}{defines whether the tibble should return qualified
file paths in the output}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-list_files"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-list_files}{}}}
\subsection{Method \code{list_files()}}{
return a vector of file paths that have been characterised
evaluates the defined private$seq_path for sequence files that are the
result of a base calling analysis using either MinKNOW or Dorado. This
method will support FASTQ, SAM and BAM file formats.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$list_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-basecalled_folder-clone"></a>}}
\if{latex}{\out{\hypertarget{method-basecalled_folder-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{basecalled_folder$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
