% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{form_arrow}
\alias{form_arrow}
\title{Prepare an arrow file from the parquet elements in current directory}
\usage{
form_arrow(dir)
}
\description{
Prepare an arrow file from the parquet elements in current directory
}
