% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_set.R
\name{sequence_set}
\alias{sequence_set}
\title{R6 Class for handling a logical set of sequence data}
\description{
\code{sequence_set} class provides an infrastructure for working with sets of
sequence that have been generated by the base calling software.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-sequence_set-new}{\code{sequence_set$new()}}
\item \href{#method-sequence_set-fastq_ingress}{\code{sequence_set$fastq_ingress()}}
\item \href{#method-sequence_set-sam_bam_ingress}{\code{sequence_set$sam_bam_ingress()}}
\item \href{#method-sequence_set-clone}{\code{sequence_set$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set-new"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set-fastq_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set-fastq_ingress}{}}}
\subsection{Method \code{fastq_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set$fastq_ingress(fastq_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set-sam_bam_ingress"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set-sam_bam_ingress}{}}}
\subsection{Method \code{sam_bam_ingress()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set$sam_bam_ingress(bam_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sequence_set-clone"></a>}}
\if{latex}{\out{\hypertarget{method-sequence_set-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{sequence_set$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
